backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
PyJWT==2.8.0
python-dotenv==1.0.0
bcrypt==4.0.1
Flask-Cors==4.0.0
